# src/net_logger/openapi.yaml
openapi: 3.0.3
info:
  title: Net Logger API
  version: 0.1.0
  description: 'Local JSON/CSV API for the Net Logger amateur-radio net logging application.


    The application is intended for trusted local or LAN use. It does not currently

    implement built-in authentication or authorization.

    '
servers:
- url: http://127.0.0.1:8088
  description: Local development server
- url: http://192.168.1.70:8088
  description: Example LAN server
...
paths:
  /api/stations:
    get:
      tags:
      - Stations
      summary: List known stations
      operationId: listStations
      parameters:
      - name: q
        in: query
        required: false
        description: Optional case-insensitive callsign/name search string.
        schema:
          type: string
          example: K5
      responses:
        '200':
          description: Stations ordered by callsign.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Station'
    post:
      tags:
      - Stations
      summary: Create a station
      operationId: createStation
      description: 'Creates a station by callsign. Callsigns are normalized to uppercase with

        whitespace removed. If the callsign already exists, the existing station is

        returned with HTTP 200.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StationCreateRequest'
            examples:
              manualStation:
                value:
                  callsign: K5SUB
                  name: Daniel Butler
                  city: Lufkin
                  state: TX
                  grid: EM21QG
                  lat: 31.27
                  lon: -94.6469
                  notes: optional
                  source: manual
      responses:
        '201':
          description: Station created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '200':
          description: Station already existed; existing record returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '400':
          description: Callsign is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stations/{station_id}:
    delete:
      tags:
      - Stations
      summary: Delete a station
      operationId: deleteStation
      description: Deletes one known station. Any check-ins for the station are removed by database cascade.
      parameters:
      - $ref: '#/components/parameters/StationId'
      responses:
        '204':
          description: Station deleted.
        '404':
          description: Station not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stations/{station_id}/refresh-fcc:
    post:
      tags:
      - Stations
      summary: Refresh station details from the local FCC database
      operationId: refreshStationFromFcc
      description: Looks up the station callsign in the configured local FCC database and updates name, city, state, grid, latitude, longitude, and source=fcc. Notes and last-heard timestamps are preserved.
      parameters:
      - $ref: '#/components/parameters/StationId'
      responses:
        '200':
          description: Station updated from FCC data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '404':
          description: Station or FCC record not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/lookup:
    get:
      tags:
      - FCC Lookup
      summary: Look up a callsign in the local FCC database
      operationId: lookupCallsign
      description: 'Uses local/off-grid FCC flat-file data. Set NET_LOGGER_FCC_LOOKUP_PATH to

        the directory containing the FCC lookup files. See docs/FCC_DATABASE_API.md

        for integration notes.

        '
      parameters:
      - name: callsign
        in: query
        required: true
        schema:
          type: string
          example: K5SUB
      responses:
        '200':
          description: Lookup result. Not-found responses still return 200 with found=false.
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/FccLookupFoundResponse'
                - $ref: '#/components/schemas/FccLookupNotFoundResponse'
...
  /api/fcc/update:
    post:
      tags:
      - FCC Lookup
      summary: Download and rebuild the local FCC database index
      operationId: updateFccDatabase
      description: 'Downloads the FCC amateur-license complete file, extracts EN.dat and HD.dat,

        and rebuilds data/EN.idx. When HD.dat is available, only active licenses are

        indexed.

        '
      responses:
        '200':
          description: FCC database update succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FccUpdateResponse'
        '500':
          description: FCC database update failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FccUpdateErrorResponse'
...
  /api/sessions/start:
    post:
      tags:
      - Sessions
      summary: Start a new net session
      operationId: startSession
      description: 'Creates a new open net session. If net_control is provided, the callsign is

        normalized, ensured as a station, and automatically checked in as sequence 1.

        '
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionStartRequest'
            example:
              name: Weekly Net
              frequency: 146.520 MHz
              net_control: K5SUB
      responses:
        '201':
          description: Session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetSession'
  /api/sessions/{session_id}/board:
    get:
      tags:
      - Sessions
      summary: Get the active board for a session
      operationId: getSessionBoard
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session metadata, known stations not checked in, and check-ins.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionBoard'
        '404':
          description: Session not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sessions/{session_id}/checkins:
    post:
      tags:
      - Check-ins
      summary: Create a check-in for a session
      operationId: createCheckin
      description: 'Adds a station to the selected session. Duplicate check-ins for the same

        session/station are idempotent and return the existing check-in with HTTP 200.

        '
      parameters:
      - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckinCreateRequest'
            example:
              station_id: 1
              traffic: true
              traffic_details: Announcement for the group
              notes: Optional signal/report notes
      responses:
        '201':
          description: Check-in created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checkin'
        '200':
          description: Check-in already existed; existing record returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checkin'
        '400':
          description: Station ID is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/checkins/{checkin_id}:
    patch:
      tags:
      - Check-ins
      summary: Update check-in traffic or notes
      operationId: updateCheckin
      parameters:
      - $ref: '#/components/parameters/CheckinId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckinUpdateRequest'
            example:
              traffic: true
              traffic_details: Updated details
              notes: Updated notes
      responses:
        '200':
          description: Updated check-in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checkin'
        '404':
          description: Check-in not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - Check-ins
      summary: Delete one check-in
      operationId: deleteCheckin
      parameters:
      - $ref: '#/components/parameters/CheckinId'
      responses:
        '204':
          description: Check-in deleted. Empty response body.
        '404':
          description: Check-in not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sessions/{session_id}/stop:
    post:
      tags:
      - Sessions
      summary: Stop a net session
      operationId: stopSession
      description: Marks the session closed and sets closed_at if it was not already closed.
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Updated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetSession'
        '404':
          description: Session not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/export.csv:
    get:
      tags:
      - Export
      summary: Export all or one session as CSV
      operationId: exportCsv
      parameters:
      - name: session_id
        in: query
        required: false
        description: Optional session ID. If omitted, all sessions are exported.
        schema:
          type: integer
          example: 12
      responses:
        '200':
          description: CSV export.
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename=net-logger-export.csv
          content:
            text/csv:
              schema:
                type: string
                example: 'session_id,session_name,frequency,net_control,status,started_at,closed_at,sequence,checked_in_at,callsign,station_name,city,state,grid,lat,lon,traffic,traffic_details,checkin_notes

                  1,Weekly Net,146.520 MHz,K5SUB,closed,2026-06-16 19:00:00,2026-06-16 20:00:00,1,2026-06-16 19:01:00,K5SUB,Daniel Butler,Lufkin,TX,EM21QG,31.27,-94.6469,no,,

                  '
  /api/records:
    delete:
      tags:
      - Administrative
      summary: Delete all application records
      operationId: deleteAllRecords
      description: 'Deletes all check-ins, net sessions, and stations from the SQLite database.

        This is destructive and cannot be undone except by restoring a backup.

        '
      parameters:
      - name: dry_run
        in: query
        required: false
        description: Return record counts without deleting anything.
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Delete result or dry-run counts.
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/DeleteRecordsResponse'
                - $ref: '#/components/schemas/DeleteRecordsDryRunResponse'
components:
  parameters:
...
    StationId:
      name: station_id
      in: path
      required: true
      schema:
        type: integer
      example: 1
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          example: session not found
    Station:
      type: object
      properties:
        id:
          type: integer
          example: 1
        callsign:
          type: string
          example: K5SUB
        name:
          type: string
          example: Daniel Butler
        city:
          type: string
          example: Lufkin
        state:
          type: string
          example: TX
        grid:
          type: string
          example: EM21QG
        lat:
          type: number
          nullable: true
          example: 31.27
        lon:
          type: number
          nullable: true
          example: -94.6469
        notes:
          type: string
          example: optional
        source:
          type: string
          example: manual
        created_at:
          type: string
          example: '2026-06-16 19:00:00'
        last_heard_at:
          type: string
          nullable: true
          example: '2026-06-16 19:01:00'
...
    FccLookupFoundResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        found:
          type: boolean
          example: true
        result:
          $ref: '#/components/schemas/FccLookupRecord'
...